+++ src/main/kotlin/cmds/Warframe.kt
@@ -57,11 +57,16 @@ object Warframe : IBase, ILogger {
                         withContent(":information_source: Assuming you meant warframe-${cmdMatches.entries.first().key}...")
                     }
                 }
-                cmdMatches.entries.first().value.handler(event)
+                cmdMatches.entries.first().value.delegateCommand(event, args)
             }
             else -> {
                 if (cmdMatches.entries.all { it.value == cmdMatches.entries.first().value }) {
-                    cmdMatches.entries.first().value.handler(event)
+                    if (args[0] != cmdMatches.entries.first().key) {
+                        buildMessage(event.channel) {
+                            withContent(":information_source: Assuming you meant warframe-${cmdMatches.entries.first().key}...")
+                        }
+                    }
+                    cmdMatches.entries.first().value.delegateCommand(event, args)
                 } else {
                     buildMessage(event.channel) {
                         withContent("Your message matches multiple commands!")