+++ src/main/kotlin/cmds/warframe/News.kt
@@ -34,11 +34,6 @@ import java.util.*
 object News : IBase {
     override fun handler(event: MessageReceivedEvent): Parser.HandleState {
         val args = getArgumentList(event.message.content).drop(1)
-        if (args.isNotEmpty() && args.any { it.matches(Regex("-{0,2}help")) }) {
-            help(event, false)
-
-            return Parser.HandleState.HANDLED
-        }
 
         buildEmbed(event.channel) {
             withTitle("Warframe News")