+++ src/main/kotlin/cmds/warframe/Ping.kt
@@ -13,11 +13,6 @@ import kotlin.system.measureTimeMillis
 object Ping : IBase, ILogger {
     override fun handler(event: MessageReceivedEvent): Parser.HandleState {
         val args = getArgumentList(event.message.content).drop(1)
-        if (args.isNotEmpty() && args.any { it.matches(Regex("-{0,2}help")) }) {
-            help(event, false)
-
-            return Parser.HandleState.HANDLED
-        }
 
         event.channel.typingStatus = true
         buildEmbed(event.channel) {
@@ -25,6 +20,7 @@ object Ping : IBase, ILogger {
 
             connections.forEach { (server, url, expectedResponse) ->
                 var responseCode = 0
+                logger.info("Pinging $server at $url...")
                 val time = measureTimeMillis {
                     val connection = URL(url).openConnection().also {
                         it.connectTimeout = 10000