+++ src/main/kotlin/cmds/warframe/Wiki.kt
@@ -31,11 +31,6 @@ import sx.blah.discord.handle.impl.events.guild.channel.message.MessageReceivedE
 object Wiki : IBase, ILogger {
     override fun handler(event: MessageReceivedEvent): Parser.HandleState {
         val args = getArgumentList(event.message.content).drop(1)
-        if (args.isNotEmpty() && args.any { it.matches(Regex("-{0,2}help")) }) {
-            help(event, false)
-
-            return Parser.HandleState.HANDLED
-        }
 
         val str = args.joinToString(" ")
                 .split(" ").joinToString(" ") {